backend-flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
pytest==8.0.0
pytest-cov==4.1.0
werkzeug==3.0.1
SQLAlchemy==2.0.25